mythxjs implementation working
