+++ lib/mythx.js
@@ -56,7 +56,9 @@ const truffle2MythXJSON = function(truffleJSON, toolId = 'truffle-security') {
         sourceList: sourceList,
         sources,
         toolId,
-        version,
+        // version for armlet, solcVersion for mythxjs
+        solcVersion: version,
+        version
     };
 };
