+++ package.json
@@ -14,7 +14,7 @@
     "mkdirp": "^0.5.1",
     "mocha": "^5.2.0",
     "multi-progress": "^2.0.0",
-    "mythxjs": "^0.5.2",
+    "mythxjs": "^1.3.1",
     "proxyquire": "^2.1.0",
     "remix-lib": ">=0.4.1",
     "scrypt": "^6.0.3",
